requirements.txt
aiosmtpd==1.4.6
aiosmtplib==3.0.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
atpublic==5.0
attrs==24.3.0
bcrypt==4.2.1
blinker==1.9.0
cached-property==2.0.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.2.0
fastapi==0.115.6
fastapi-mail==1.4.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jinja2==3.1.5
jose==1.0.0
jwt==1.3.1
mako==1.3.9
markupsafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-settings==2.7.0
pyjwt==2.10.1
pytest==8.3.4
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
ruff==0.8.5
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.38
sqlalchemy-orm==1.2.10
starlette==0.41.3
typing-extensions==4.12.2
typing-inspect==0.9.0
uvicorn==0.34.0
